requirements.txt
pandas
numpy
spacy
textblob
nltk
better-profanity
scikit-learn
python-dotenv